Restructure of repository. Restructure of MultiRepository.
