+++ org.eclipse.winery.common/src/main/java/org/eclipse/winery/common/exceptions/QNameAlreadyExistsException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013 Contributors to the Eclipse Foundation
+ * Copyright (c) 2019 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,7 +11,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-package org.eclipse.winery.common.interfaces;
+package org.eclipse.winery.common.exceptions;
 
 public class QNameAlreadyExistsException extends Exception {
