+++ org.eclipse.winery.frontends/app/topologymodeler/src/app/canvas/canvas.component.html
@@ -170,7 +170,7 @@
                 <button type="button" id="updateCapabilitiesBtn" (click)="updateCaps(); capForm.reset()"
                         class="btn btn-primary" [disabled]="!capForm.form.valid || duplicateId">Update
                 </button>
-                <button  (click)="deleteCapability(); capForm.reset()"
+                <button (click)="deleteCapability(); capForm.reset()"
                         type="button"
                         id="deleteCapbtn"
                         class="btn btn-primary">Delete
