+++ org.eclipse.winery.frontends/app/topologymodeler/src/app/models/requirementModel.ts
@@ -29,7 +29,6 @@ export class RequirementModel {
     public node?: string;
     public relationship?: string;
 
-
     static fromRequirementDefinition(def: RequirementDefinitionModel): RequirementModel {
         const result = new RequirementModel();
         result.any = def.any;
