+++ org.eclipse.winery.frontends/app/tosca-management/src/app/wineryReadmeModule/wineryReadme.component.html
@@ -81,7 +81,8 @@
         </tabset>
     </div>
 
-    <textarea style="width: 95%; height: 500px; resize: none; font-family: monospace" [(ngModel)]="readmeContent"></textarea>
+    <textarea style="width: 95%; height: 500px; resize: none; font-family: monospace"
+              [(ngModel)]="readmeContent"></textarea>
 </div>
