+++ org.eclipse.winery.frontends/app/tosca-management/src/app/wineryReadmeModule/wineryReadme.module.ts
@@ -11,14 +11,14 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-import {NgModule} from '@angular/core';
-import {WineryReadmeComponent} from './wineryReadme.component';
-import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
-import {WineryLoaderModule} from '../wineryLoader/wineryLoader.module';
-import {TabsModule} from 'ngx-bootstrap';
-import {WineryPipesModule} from '../wineryPipes/wineryPipes.module';
-import {WineryMarkdownModule} from '../wineryMarkdownComponent/wineryMarkdown.module';
+import { NgModule } from '@angular/core';
+import { WineryReadmeComponent } from './wineryReadme.component';
+import { CommonModule } from '@angular/common';
+import { FormsModule } from '@angular/forms';
+import { WineryLoaderModule } from '../wineryLoader/wineryLoader.module';
+import { TabsModule } from 'ngx-bootstrap';
+import { WineryPipesModule } from '../wineryPipes/wineryPipes.module';
+import { WineryMarkdownModule } from '../wineryMarkdownComponent/wineryMarkdown.module';
 
 @NgModule({
     imports: [
