+++ org.eclipse.winery.repository.client/src/main/java/org/eclipse/winery/repository/client/IWineryRepositoryClient.java
@@ -14,15 +14,13 @@
 
 package org.eclipse.winery.repository.client;
 
-import org.eclipse.winery.common.interfaces.IWineryRepository;
+import org.eclipse.winery.repository.backend.IWineryRepository;
 
 public interface IWineryRepositoryClient extends IWineryRepository {
-
     /**
      * Adds an URI to the list of known repositories
      * <p>
-     * SIDE EFFECT: If currently no primary repository is defined, the given
-     * repository is used as primary repository
+     * SIDE EFFECT: If currently no primary repository is defined, the given repository is used as primary repository
      *
      * @param uri the URI of the repository
      */
@@ -36,22 +34,18 @@ public interface IWineryRepositoryClient extends IWineryRepository {
     /**
      * Sets the primary repository
      * <p>
-     * SIDE EFFECT: If the repository is not known as general repository (via
-     * addRepository), the given repository is added to the list of known
-     * repositories
-     *
-     * @param uri
+     * SIDE EFFECT: If the repository is not known as general repository (via addRepository), the given repository is
+     * added to the list of known repositories
      */
     void setPrimaryRepository(String uri);
 
     /**
-     * Checks whether the primary repository is available to be used. Typically,
-     * this method should return "true". In case of network or server failures,
-     * the result is "false". Note that the availability may change over time
-     * and also a repository might become unavailable during querying it.
+     * Checks whether the primary repository is available to be used. Typically, this method should return "true". In
+     * case of network or server failures, the result is "false". Note that the availability may change over time and
+     * also a repository might become unavailable during querying it.
      * <p>
-     * This method also returns "false" if no primary repository is set. For
-     * instance, this is the case of no repository is registered at the client.
+     * This method also returns "false" if no primary repository is set. For instance, this is the case of no repository
+     * is registered at the client.
      *
      * @return true if the repository is reachable over network, false otherwise
      */
