+++ org.eclipse.winery.repository.client/src/test/java/org/eclipse/winery/repository/client/TestWineryRepositoryClient.java
@@ -28,7 +28,7 @@ import org.eclipse.winery.model.tosca.TTopologyTemplate;
 import org.eclipse.winery.repository.TestWithGitBackedRepository;
 import org.eclipse.winery.repository.backend.IRepository;
 import org.eclipse.winery.repository.backend.RepositoryFactory;
-import org.eclipse.winery.repository.backend.filebased.FilebasedRepository;
+import org.eclipse.winery.repository.backend.filebased.AbstractFileBasedRepository;
 import org.eclipse.winery.repository.rest.server.WineryUsingHttpServer;
 
 import org.eclipse.jdt.annotation.Nullable;
@@ -73,8 +73,8 @@ public class TestWineryRepositoryClient extends TestWithGitBackedRepository {
         server.start();
 
         IRepository repository = RepositoryFactory.getRepository();
-        if (repository instanceof FilebasedRepository) {
-            LOGGER.debug("Using path " + ((FilebasedRepository) repository).getRepositoryRoot());
+        if (repository instanceof AbstractFileBasedRepository) {
+            LOGGER.debug("Using path " + repository.getRepositoryRoot());
         } else {
             LOGGER.debug("Repository is not filebased");
         }
