+++ org.eclipse.winery.repository/src/main/java/org/eclipse/winery/repository/backend/IWineryRepository.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013-2018 Contributors to the Eclipse Foundation
+ * Copyright (c) 2019 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,7 +11,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-package org.eclipse.winery.common.interfaces;
+package org.eclipse.winery.repository.backend;
 
 import java.util.Collection;
 import java.util.List;
@@ -19,8 +19,10 @@ import java.util.SortedSet;
 
 import javax.xml.namespace.QName;
 
+import org.eclipse.winery.common.exceptions.QNameAlreadyExistsException;
 import org.eclipse.winery.common.ids.GenericId;
 import org.eclipse.winery.common.ids.definitions.DefinitionsChildId;
+import org.eclipse.winery.common.interfaces.QNameWithName;
 import org.eclipse.winery.model.tosca.TDefinitions;
 import org.eclipse.winery.model.tosca.TEntityType;
 import org.eclipse.winery.model.tosca.TExtensibleElements;
