+++ org.eclipse.winery.repository/src/main/java/org/eclipse/winery/repository/backend/RepositoryFactory.java
@@ -13,22 +13,22 @@
  *******************************************************************************/
 package org.eclipse.winery.repository.backend;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
+import java.nio.file.Paths;
 import java.util.Objects;
-import java.util.Optional;
 
 import org.eclipse.winery.common.configuration.Environments;
 import org.eclipse.winery.common.configuration.FileBasedRepositoryConfiguration;
 import org.eclipse.winery.common.configuration.GitBasedRepositoryConfiguration;
 import org.eclipse.winery.common.configuration.RepositoryConfigurationObject;
 import org.eclipse.winery.repository.backend.constants.Filename;
-import org.eclipse.winery.repository.backend.filebased.FilebasedRepository;
+import org.eclipse.winery.repository.backend.filebased.AbstractFileBasedRepository;
 import org.eclipse.winery.repository.backend.filebased.GitBasedRepository;
 import org.eclipse.winery.repository.backend.filebased.MultiRepository;
+import org.eclipse.winery.repository.backend.filebased.XMLRepository;
 import org.eclipse.winery.repository.backend.filebased.YamlRepository;
 
 import org.eclipse.jgit.api.errors.GitAPIException;
@@ -36,8 +36,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class RepositoryFactory {
-
-    public static List<FilebasedRepository> repositoryList = new ArrayList<>();
 
     private static final Logger LOGGER = LoggerFactory.getLogger(RepositoryFactory.class);
 
@@ -47,42 +45,56 @@ public class RepositoryFactory {
     private static IRepository repository = null;
 
     private static boolean repositoryContainsRepoConfig(FileBasedRepositoryConfiguration config) {
-        return FilebasedRepository.getRepositoryRoot(config).resolve(Filename.FILENAME_JSON_REPOSITORIES).toFile().exists();
+        if (config.getRepositoryPath().isPresent()) {
+            return new File(config.getRepositoryPath().get().toString(), Filename.FILENAME_JSON_REPOSITORIES).exists();
+        } else {
+            return new File(Environments.getRepositoryConfig().getRepositoryRoot(), Filename.FILENAME_JSON_REPOSITORIES).exists();
+        }
     }
 
-    public static FilebasedRepository createXmlOrYamlRepository(FileBasedRepositoryConfiguration configuration) {
+    public static AbstractFileBasedRepository createXmlOrYamlRepository(FileBasedRepositoryConfiguration configuration, Path repositoryRoot) {
         if (RepositoryConfigurationObject.RepositoryProvider.YAML.equals(configuration.getRepositoryProvider())) {
-            return new YamlRepository(configuration);
+            return new YamlRepository(repositoryRoot);
         } else {
-            // XML-based repository
-            return new FilebasedRepository(configuration);
+            return new XMLRepository(repositoryRoot);
         }
     }
 
     public static void reconfigure(GitBasedRepositoryConfiguration gitBasedRepositoryConfiguration) throws IOException, GitAPIException {
         RepositoryFactory.gitBasedRepositoryConfiguration = gitBasedRepositoryConfiguration;
         RepositoryFactory.fileBasedRepositoryConfiguration = null;
-
-        if (repositoryContainsRepoConfig(gitBasedRepositoryConfiguration)) {
-            repository = new MultiRepository(gitBasedRepositoryConfiguration, createXmlOrYamlRepository(gitBasedRepositoryConfiguration));
+        AbstractFileBasedRepository compositeRepository = null;
+        //If a repository root is specified by the configuration use it instead of the root specified in the Configuration
+        if (gitBasedRepositoryConfiguration.getRepositoryPath().isPresent()) {
+            compositeRepository = createXmlOrYamlRepository(gitBasedRepositoryConfiguration, gitBasedRepositoryConfiguration.getRepositoryPath().get());
         } else {
-            repository = new GitBasedRepository(gitBasedRepositoryConfiguration, createXmlOrYamlRepository(gitBasedRepositoryConfiguration));
+            compositeRepository = createXmlOrYamlRepository(fileBasedRepositoryConfiguration, Paths.get(Environments.getRepositoryConfig().getRepositoryRoot()));
+        }
+        if (repositoryContainsRepoConfig(gitBasedRepositoryConfiguration)) {
+            repository = new MultiRepository(compositeRepository.getRepositoryRoot());
+        } else {
+            repository = new GitBasedRepository(gitBasedRepositoryConfiguration, compositeRepository);
         }
     }
 
     public static void reconfigure(FileBasedRepositoryConfiguration fileBasedRepositoryConfiguration) {
         RepositoryFactory.fileBasedRepositoryConfiguration = fileBasedRepositoryConfiguration;
         RepositoryFactory.gitBasedRepositoryConfiguration = null;
-
+        AbstractFileBasedRepository compositeRepository;
+        //If a repository root is specified by the configuration use it instead of the root specified in the Configuration
+        if (fileBasedRepositoryConfiguration.getRepositoryPath().isPresent()) {
+            compositeRepository = createXmlOrYamlRepository(fileBasedRepositoryConfiguration, fileBasedRepositoryConfiguration.getRepositoryPath().get());
+        } else {
+            compositeRepository = createXmlOrYamlRepository(fileBasedRepositoryConfiguration, Paths.get(Environments.getRepositoryConfig().getRepositoryRoot()));
+        }
         if (repositoryContainsRepoConfig(fileBasedRepositoryConfiguration)) {
             try {
-                FilebasedRepository compositeRepository = createXmlOrYamlRepository(fileBasedRepositoryConfiguration);
-                repository = new MultiRepository(new GitBasedRepositoryConfiguration(false, fileBasedRepositoryConfiguration), compositeRepository);
+                repository = new MultiRepository(compositeRepository.getRepositoryRoot());
             } catch (IOException | GitAPIException exception) {
                 exception.printStackTrace();
             }
         } else {
-            repository = new FilebasedRepository(fileBasedRepositoryConfiguration);
+            repository = compositeRepository;
         }
     }
 
@@ -90,19 +102,16 @@ public class RepositoryFactory {
      * Reconfigures based on Environment
      */
     public static void reconfigure() throws Exception {
-        final Optional<GitBasedRepositoryConfiguration> gitBasedRepositoryConfiguration = Environments.getGitBasedRepositoryConfiguration();
+        final GitBasedRepositoryConfiguration gitBasedRepositoryConfiguration = Environments.getGitBasedRepositoryConfiguration();
         final FileBasedRepositoryConfiguration filebasedRepositoryConfiguration = Environments.getFilebasedRepositoryConfiguration();
 
         // Determine whether the filebased repository could be git repository.
         // We do not use JGit's capabilities, but do it just by checking for the existance of a ".git" directory.
-        final Path repositoryRoot = FilebasedRepository.getRepositoryRoot(filebasedRepositoryConfiguration);
+        final Path repositoryRoot = Paths.get(Environments.getRepositoryConfig().getRepositoryRoot());
         final Path gitDirectory = repositoryRoot.resolve(".git");
         boolean isGit = (Files.exists(gitDirectory) && Files.isDirectory(gitDirectory));
-
-        if (gitBasedRepositoryConfiguration.isPresent()) {
-            reconfigure(gitBasedRepositoryConfiguration.get());
-        } else if (isGit) {
-            reconfigure(new GitBasedRepositoryConfiguration(false, filebasedRepositoryConfiguration));
+        if (isGit) {
+            reconfigure(gitBasedRepositoryConfiguration);
         } else {
             reconfigure(filebasedRepositoryConfiguration);
         }
@@ -126,10 +135,10 @@ public class RepositoryFactory {
         return getRepository(fileBasedRepositoryConfiguration);
     }
 
-    public static IRepository getRepository(FileBasedRepositoryConfiguration fileBasedRepositoryConfiguration) {
+    public static AbstractFileBasedRepository getRepository(FileBasedRepositoryConfiguration fileBasedRepositoryConfiguration) {
         // FIXME: currently, the CSAR export does not reuse the repository instance returned here. Thus, we have to reconfigure the repository.
         // This should be fixed by always passing IRepository when working with the repository
         reconfigure(fileBasedRepositoryConfiguration);
-        return createXmlOrYamlRepository(fileBasedRepositoryConfiguration);
+        return createXmlOrYamlRepository(fileBasedRepositoryConfiguration, Paths.get(Environments.getRepositoryConfig().getRepositoryRoot()));
     }
 }
