+++ org.eclipse.winery.repository/src/main/java/org/eclipse/winery/repository/backend/filebased/MultiRepository.java
@@ -19,12 +19,16 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.FileTime;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedSet;
@@ -32,16 +36,31 @@ import java.util.TreeSet;
 
 import org.eclipse.winery.common.Constants;
 import org.eclipse.winery.common.RepositoryFileReference;
+import org.eclipse.winery.common.configuration.Environments;
 import org.eclipse.winery.common.configuration.FileBasedRepositoryConfiguration;
 import org.eclipse.winery.common.configuration.GitBasedRepositoryConfiguration;
 import org.eclipse.winery.common.ids.GenericId;
 import org.eclipse.winery.common.ids.Namespace;
+import org.eclipse.winery.common.ids.admin.AccountabilityId;
+import org.eclipse.winery.common.ids.admin.EdmmMappingsId;
 import org.eclipse.winery.common.ids.definitions.DefinitionsChildId;
 import org.eclipse.winery.common.ids.elements.ToscaElementId;
+import org.eclipse.winery.model.tosca.Definitions;
+import org.eclipse.winery.repository.backend.AccountabilityConfigurationManager;
+import org.eclipse.winery.repository.backend.BackendUtils;
+import org.eclipse.winery.repository.backend.EdmmManager;
+import org.eclipse.winery.repository.backend.IRepository;
 import org.eclipse.winery.repository.backend.NamespaceManager;
 import org.eclipse.winery.repository.backend.RepositoryFactory;
+import org.eclipse.winery.repository.backend.constants.Filename;
+import org.eclipse.winery.repository.backend.filebased.management.IRepositoryResolver;
+import org.eclipse.winery.repository.backend.filebased.management.RepositoryResolverFactory;
 import org.eclipse.winery.repository.exceptions.WineryRepositoryException;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectReader;
+import com.fasterxml.jackson.databind.SerializationFeature;
 import org.apache.commons.configuration2.Configuration;
 import org.apache.tika.mime.MediaType;
 import org.eclipse.jgit.api.errors.GitAPIException;
@@ -51,63 +70,63 @@ import org.slf4j.LoggerFactory;
 /**
  * Layer that manages the local repositories
  */
-public class MultiRepository extends GitBasedRepository {
+public class MultiRepository implements IRepository {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MultiRepository.class);
 
-    private final Map<FilebasedRepository, Set<String>> repositoryGlobal = new HashMap<>();
-    private final Map<FilebasedRepository, Set<Namespace>> repositoryCommonNamespace = new HashMap<>();
-
+    private final Map<IRepository, Set<String>> repositoryGlobal = new HashMap<>();
+    private final Map<IRepository, Set<Namespace>> repositoryCommonNamespace = new HashMap<>();
+    private File repositoryConfiguration;
+    private List<RepositoryProperties> repositoriesList = new ArrayList<>();
     private GitBasedRepository localRepository;
+    private Path repositoryRoot;
 
-    public MultiRepository(GitBasedRepositoryConfiguration configuration, FilebasedRepository repository) throws IOException, GitAPIException {
-        super(configuration, repository);
-
+    /**
+     * Initializes a local Git Repository at repositoryRoot/workspace. The repository root points to the directory of
+     * the MultiRepository. The repositoryConfiguration points to the repositories.json file. This file contains the
+     * repositories that are part of the MultiRepository.
+     */
+    public MultiRepository(Path repositoryRoot) throws IOException, GitAPIException {
+        this.repositoryRoot = repositoryRoot;
         try {
             LOGGER.debug("Trying to initialize local repository...");
-            File localRepoPath = new File(FilebasedRepository.getActiveRepositoryFilePath(), Constants.DEFAULT_LOCAL_REPO_NAME);
+            File localRepoPath = new File(repositoryRoot.toString(), Constants.DEFAULT_LOCAL_REPO_NAME);
             FileBasedRepositoryConfiguration localRepoConfig = new FileBasedRepositoryConfiguration(localRepoPath.toPath());
             GitBasedRepositoryConfiguration gitConfig = new GitBasedRepositoryConfiguration(false, localRepoConfig);
-            this.localRepository = new GitBasedRepository(gitConfig, RepositoryFactory.createXmlOrYamlRepository(localRepoConfig));
+            this.localRepository = new GitBasedRepository(gitConfig, RepositoryFactory.createXmlOrYamlRepository(localRepoConfig, localRepoPath.toPath()));
             LOGGER.debug("Local repo has been initialized at {}", localRepoPath.getAbsolutePath());
+            repositoryConfiguration = new File(this.getRepositoryRoot().toString(), Filename.FILENAME_JSON_REPOSITORIES);
         } catch (IOException | GitAPIException e) {
             LOGGER.error("Error while initializing local repository of the Multi Repository!", e);
             throw e;
         }
-
-        RepositoryConfigurationManager.initialize(this);
-        RepositoryUtils.checkGitIgnore(this);
+        readRepositoriesConfig();
         repositoryGlobal.put(localRepository, new HashSet<>());
         updateNamespaces();
     }
 
-    @Override
-    Path generateWorkingRepositoryRoot() {
-        return this.getRepositoryDep();
-    }
-
-    protected FilebasedRepository getLocalRepository() {
+    IRepository getLocalRepository() {
         return localRepository;
     }
 
-    protected Map<FilebasedRepository, Set<String>> getRepositoriesMap() {
+    Map<IRepository, Set<String>> getRepositoriesMap() {
         return repositoryGlobal;
     }
 
-    protected Map<FilebasedRepository, Set<Namespace>> getRepositoriesCommonNamespace() {
+    Map<IRepository, Set<Namespace>> getRepositoriesCommonNamespace() {
         return repositoryCommonNamespace;
     }
 
-    protected Collection<FilebasedRepository> getRepositories() {
+    public Set<IRepository> getRepositories() {
         return repositoryGlobal.keySet();
     }
 
-    protected void addRepository(FilebasedRepository repository) {
+    private void addRepository(IRepository repository) {
         registerRepository(repository);
     }
 
     protected void removeRepository(String urlToRepository) {
-        for (FilebasedRepository repo : repositoryGlobal.keySet()) {
+        for (IRepository repo : repositoryGlobal.keySet()) {
             if (((GitBasedRepository) repo).getRepositoryUrl().equals(urlToRepository)) {
                 unregisterRepository(repo);
                 break;
@@ -115,11 +134,11 @@ public class MultiRepository extends GitBasedRepository {
         }
     }
 
-    protected void removeRepository(FilebasedRepository repository) {
+    protected void removeRepository(XMLRepository repository) {
         unregisterRepository(repository);
     }
 
-    private void addNamespacesToRepository(FilebasedRepository repository, GenericId id) {
+    private void addNamespacesToRepository(IRepository repository, GenericId id) {
         if (id instanceof DefinitionsChildId) {
             Namespace namespace = ((DefinitionsChildId) id).getNamespace();
             String ns = namespace.getDecoded();
@@ -143,12 +162,12 @@ public class MultiRepository extends GitBasedRepository {
         }
     }
 
-    private void addNamespacesToRepository(FilebasedRepository repository, RepositoryFileReference ref) {
+    private void addNamespacesToRepository(IRepository repository, RepositoryFileReference ref) {
         addNamespacesToRepository(repository, ref.getParent());
     }
 
-    protected void updateNamespaces() {
-        Map<FilebasedRepository, Set<String>> tempMap = new HashMap<>(repositoryGlobal);
+    void updateNamespaces() {
+        Map<IRepository, Set<String>> tempMap = new HashMap<>(repositoryGlobal);
         tempMap.keySet().forEach(repository -> {
             Collection<String> repositoryNamespaces = repository.getNamespaceManager().getAllNamespaces().keySet();
             repositoryGlobal.put(repository, new HashSet<>(repositoryNamespaces));
@@ -173,7 +192,7 @@ public class MultiRepository extends GitBasedRepository {
         return setNS;
     }
 
-    private void registerRepository(FilebasedRepository repository) {
+    private void registerRepository(IRepository repository) {
         if (repositoryGlobal.get(repository) != null) {
             LOGGER.debug("The repository is probably already registered.");
             return;
@@ -183,9 +202,148 @@ public class MultiRepository extends GitBasedRepository {
         updateNamespaces();
     }
 
-    private void unregisterRepository(FilebasedRepository repository) {
+    private void unregisterRepository(IRepository repository) {
         repositoryGlobal.remove(repository);
         repositoryCommonNamespace.remove(repository);
+    }
+
+    /**
+     * Clones the repositories specified by repositories into the MultiRepository
+     *
+     * @param repositories the set of repoistories that should be cloned into the MultiRepository.
+     */
+    void addRepositoryToFile(List<RepositoryProperties> repositories) {
+        repositoriesList = repositories;
+        saveConfiguration();
+        loadRepositoriesByList();
+    }
+
+    /**
+     * Returns the repositoryList, if the repositories.json exist, the repositoriesList is first
+     * loaded with the repositories from the json.
+     */
+    List<RepositoryProperties> getRepositoriesFromFile() {
+        if (repoContainsConfigFile()) {
+            try {
+                readRepositoriesConfig();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+        return repositoriesList;
+    }
+
+    /**
+     * Loads the repositories from the repositories.json file 
+     * into the repositories list.
+     * Then clones the repositories from the repositories list.
+     * @throws IOException
+     */
+    private void readRepositoriesConfig() throws IOException {
+        if (repoContainsConfigFile()) {
+            LOGGER.info("Found Repositories file");
+            loadConfiguration(repositoryConfiguration);
+            loadRepositoriesByList();
+        } else {
+            createConfigFileAndSetFactoryToMultiRepository();
+        }
+    }
+
+    /**
+     * Writes the content of the repositoriesList into the repositories.json of the MultiRepository.
+     */
+    private void saveConfiguration() {
+        if (!repoContainsConfigFile()) {
+            createConfigFileAndSetFactoryToMultiRepository();
+        } else {
+            ObjectMapper objectMapper = new ObjectMapper();
+            try {
+                objectMapper.enable(SerializationFeature.INDENT_OUTPUT);
+                objectMapper.writeValue(repositoryConfiguration, repositoriesList);
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    /**
+     * @return True whenever there is a repositories.json file in the root folder of the MultiRepository
+     */
+    private boolean repoContainsConfigFile() {
+        File repo = new File(this.getRepositoryRoot().toString(), Filename.FILENAME_JSON_REPOSITORIES);
+        return repo.exists();
+    }
+
+    /**
+     * Reads the dependencies file into the repositoriesList.
+     *
+     * @param dependency The path to the dependencies file
+     */
+    private void loadConfiguration(File dependency) throws IOException {
+        ObjectMapper objectMapper = new ObjectMapper();
+        ObjectReader reader = objectMapper.readerFor(new TypeReference<List<RepositoryProperties>>() {
+        });
+        repositoriesList = reader.readValue(dependency);
+    }
+
+    /**
+     * Clones all repositories of the MultiRepository into the file system.
+     * Does not clone duplicates. Also clones all dependencies of the repositories
+     */
+    private void loadRepositoriesByList() {
+        for (RepositoryProperties repository : repositoriesList) {
+            createRepository(repository.getUrl(), repository.getBranch());
+        }
+    }
+
+    private void createConfigFileAndSetFactoryToMultiRepository() {
+        MultiRepositoryManager.initializeRepositoryListForMultiRepositoryAndReconfigureFactory(repositoriesList);
+        try {
+            RepositoryFactory.reconfigure();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * This method clones a repository into the file system.
+     * If the cloned repository contains dependencies in the form of a repositories.json file,
+     * the dependencies will be cloned recursively if they are not already in the MultiRepository.
+     * The subrepositories are GitbasedRepositories and are added to the list of repositories to the MultiRepository. 
+     * It the subrepositories have dependencies, they are initialized as MultiRepos
+     * @param url of the repository
+     * @param branch which should be cloned
+     */
+    private void createRepository(String url, String branch) {
+        IRepositoryResolver resolver = null;
+        if (RepositoryResolverFactory.getResolver(url, branch).isPresent()) {
+            resolver = RepositoryResolverFactory.getResolver(url, branch).get();
+        }
+
+        if (resolver != null && !RepositoryUtils.checkRepositoryDuplicate(url, this)) {
+            String ownerDirectory;
+            File ownerRootFile;
+            try {
+                ownerDirectory = URLEncoder.encode(resolver.getRepositoryMaintainerUrl(), "UTF-8");
+                ownerRootFile = new File(Environments.getRepositoryConfig().getRepositoryRoot(), ownerDirectory);
+                if (!ownerRootFile.exists()) {
+                    Files.createDirectories(ownerRootFile.toPath());
+                }
+
+                File repositoryLocation = new File(ownerRootFile, resolver.getRepositoryName());
+                IRepository newSubRepository = resolver.createRepository(repositoryLocation);
+                this.addRepository(newSubRepository);
+
+                File configurationFile = new File(newSubRepository.getRepositoryRoot().toString().replace("\\workspace", ""), Filename.FILENAME_JSON_REPOSITORIES);
+                if (configurationFile.exists()) {
+                    loadConfiguration(configurationFile);
+                    loadRepositoriesByList();
+                }
+            } catch (IOException | GitAPIException e) {
+                LOGGER.error("Error while creating the repository structure");
+                e.printStackTrace();
+            }
+        }
     }
 
     @Override
@@ -219,6 +377,37 @@ public class MultiRepository extends GitBasedRepository {
     }
 
     @Override
+    public Definitions definitionsFromRef(RepositoryFileReference ref) throws IOException {
+        return localRepository.definitionsFromRef(ref);
+    }
+
+    @Override
+    public void doDump(OutputStream out) throws IOException {
+        localRepository.doDump(out);
+    }
+
+    @Override
+    public void doClear() {
+        localRepository.doClear();
+    }
+
+    @Override
+    public EdmmManager getEdmmManager() {
+        RepositoryFileReference ref = BackendUtils.getRefOfJsonConfiguration(new EdmmMappingsId());
+        return new JsonBasedEdmmManager(ref2AbsolutePath(ref).toFile());
+    }
+
+    @Override
+    public AccountabilityConfigurationManager getAccountabilityConfigurationManager() {
+        RepositoryFileReference repoRef = BackendUtils.getRefOfConfiguration(new AccountabilityId());
+        RepositoryFileReference keystoreRef = new RepositoryFileReference(new AccountabilityId(), "CustomKeystore.json");
+        RepositoryFileReference defaultKeystoreRef = new RepositoryFileReference(new AccountabilityId(), "DefaultKeystore.json");
+
+        return AccountabilityConfigurationManager.getInstance(ref2AbsolutePath(repoRef).toFile(),
+            ref2AbsolutePath(keystoreRef).toFile(), ref2AbsolutePath(defaultKeystoreRef).toFile());
+    }
+
+    @Override
     public boolean flagAsExisting(GenericId id) {
         return RepositoryUtils.getRepositoryById(id, this).flagAsExisting(id);
     }
@@ -230,14 +419,14 @@ public class MultiRepository extends GitBasedRepository {
 
     @Override
     public void putContentToFile(RepositoryFileReference ref, String content, MediaType mediaType) throws IOException {
-        FilebasedRepository repository = RepositoryUtils.getRepositoryByRef(ref, this);
+        IRepository repository = RepositoryUtils.getRepositoryByRef(ref, this);
         repository.putContentToFile(ref, content, mediaType);
         addNamespacesToRepository(repository, ref);
     }
 
     @Override
     public void putContentToFile(RepositoryFileReference ref, InputStream inputStream, MediaType mediaType) throws IOException {
-        FilebasedRepository repository = RepositoryUtils.getRepositoryByRef(ref, this);
+        IRepository repository = RepositoryUtils.getRepositoryByRef(ref, this);
         repository.putContentToFile(ref, inputStream, mediaType);
         addNamespacesToRepository(repository, ref);
     }
@@ -269,6 +458,31 @@ public class MultiRepository extends GitBasedRepository {
     }
 
     @Override
+    public Path ref2AbsolutePath(RepositoryFileReference ref) {
+        return null;
+    }
+
+    @Override
+    public Path id2RelativePath(GenericId id) {
+        return null;
+    }
+
+    @Override
+    public Path id2AbsolutePath(GenericId id) {
+        return null;
+    }
+
+    @Override
+    public Path makeAbsolute(Path relativePath) {
+        return null;
+    }
+
+    @Override
+    public Path getRepositoryRoot() {
+        return this.repositoryRoot;
+    }
+
+    @Override
     public <T extends DefinitionsChildId> SortedSet<T> getAllDefinitionsChildIds(Class<T> idClass) {
         SortedSet<T> result = new TreeSet<>();
         getRepositories().forEach(repository -> result.addAll(repository.getAllDefinitionsChildIds(idClass)));
@@ -294,12 +508,22 @@ public class MultiRepository extends GitBasedRepository {
 
     @Override
     public void forceDelete(GenericId id) {
-        RepositoryUtils.getRepositoryById(id, this).forceDelete(id);
+        try {
+            RepositoryUtils.getRepositoryById(id, this).forceDelete(id);
+        } catch (IOException ioex) {
+            LOGGER.debug("Error while force deleting definition child.", ioex);
+        }
     }
 
     @Override
     public void forceDelete(Class<? extends DefinitionsChildId> definitionsChildIdClazz, Namespace namespace) {
-        getRepositories().forEach(repository -> repository.forceDelete(definitionsChildIdClazz, namespace));
+        getRepositories().forEach(repository -> {
+            try {
+                repository.forceDelete(definitionsChildIdClazz, namespace);
+            } catch (IOException ioex) {
+                LOGGER.debug("Error while force deleting definition child.", ioex);
+            }
+        });
     }
 
     @Override
