+++ org.eclipse.winery.repository/src/test/java/org/eclipse/winery/repository/backend/BackendUtilsTest.java
@@ -22,6 +22,8 @@ import java.util.Optional;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
+import javax.xml.namespace.QName;
+
 import org.eclipse.winery.common.RepositoryFileReference;
 import org.eclipse.winery.common.ids.definitions.ArtifactTemplateId;
 import org.eclipse.winery.model.tosca.TArtifactReference;
@@ -36,8 +38,6 @@ import org.eclipse.winery.repository.datatypes.ids.elements.ArtifactTemplateSour
 import org.junit.jupiter.api.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 import org.xmlunit.matchers.CompareMatcher;
-
-import javax.xml.namespace.QName;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
@@ -194,7 +194,7 @@ public class BackendUtilsTest {
         // alternative test implementation: Use git-based repository
         // this test at hand is closer to the implementation, but easier to write
 
-        IGenericRepository repository = mock(IGenericRepository.class);
+        IRepository repository = mock(IRepository.class);
         ArtifactTemplateFilesDirectoryId artifactTemplateFilesDirectoryId = new ArtifactTemplateFilesDirectoryId(artifactTemplateId);
         when(repository.getContainedFiles(artifactTemplateFilesDirectoryId)).thenReturn(Collections.emptySortedSet());
 
@@ -212,7 +212,7 @@ public class BackendUtilsTest {
         // alternative test implementation: Use git-based repository
         // this test at hand is closer to the implementation, but easier to write
 
-        IGenericRepository repository = mock(IGenericRepository.class);
+        IRepository repository = mock(IRepository.class);
         ArtifactTemplateFilesDirectoryId artifactTemplateFilesDirectoryId = new ArtifactTemplateFilesDirectoryId(artifactTemplateId);
         when(repository.getContainedFiles(artifactTemplateFilesDirectoryId)).thenReturn(Collections.emptySortedSet());
 
@@ -230,7 +230,7 @@ public class BackendUtilsTest {
         // alternative test implementation: Use git-based repository
         // this test at hand is closer to the implementation, but easier to write
 
-        IGenericRepository repository = mock(IGenericRepository.class);
+        IRepository repository = mock(IRepository.class);
         ArtifactTemplateFilesDirectoryId artifactTemplateFilesDirectoryId = new ArtifactTemplateFilesDirectoryId(artifactTemplateId);
 
         SortedSet<RepositoryFileReference> containedReferences = new TreeSet<>();
